move all files at package model to package models in test folder, delete ExchangedDocumentTest and IncludesTest at package models in main folder
